% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_gradient.R
\name{apply_gradient2}
\alias{apply_gradient2}
\title{Compute gradient contribution for exact response}
\usage{
apply_gradient2(
  x_train,
  y_train,
  model,
  img_train = NULL,
  verbose = FALSE,
  response_varying = FALSE,
  lambda2 = 0,
  numnet = 1
)
}
\description{
Compute gradient contribution for exact response
}
\examples{
mo <- ontram_polr(x_dim = 1L, y_dim = 5L, method = "logit")
x_train <- matrix(c(0.5, -0.5, 0), ncol = 1)
x_train <- tf$constant(x_train, dtype = "float32")
y_train <- matrix(c(0, 1, 0, 0, 0,
                    0, 0, 1, 0, 0,
                    0, 0, 0, 0, 1), nrow = 3, byrow = TRUE)
y_train <- tf$constant(y_train, dtype = "float32")
apply_gradient(x_train, y_train, mo, verbose = TRUE)
}
