% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram_methods.R
\name{coef.k_ontram}
\alias{coef.k_ontram}
\title{S3 methods for \code{k_ontram}}
\usage{
\method{coef}{k_ontram}(object, which = NULL, ...)
}
\arguments{
\item{object}{object of class \code{k_ontram}}

\item{which}{optional name of layer for which to return coefficients}
}
\description{
S3 methods for \code{k_ontram}
}
