% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ontram}
\alias{coef.ontram}
\title{Function for coef for simple ontram}
\usage{
\method{coef}{ontram}(object, with_baseline = FALSE, ...)
}
\description{
Function for coef for simple ontram
}
\examples{
mod <- ontram_polr(2L, 5L)
coef(mod, with_baseline = TRUE)
}
