% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_logLik.R
\name{compute_logLik2}
\alias{compute_logLik2}
\title{Compute logLik contributions for exact response}
\usage{
compute_logLik2(
  gammas,
  beta = NULL,
  eta = NULL,
  y_train,
  distr = k_sigmoid,
  lambda2 = 0,
  numnet = 1
)
}
\description{
Compute logLik contributions for exact response
}
\examples{
gammas <- matrix(c(-1, -1, -1, -1, -2, -1, -1, -1), nrow = 2, byrow = TRUE)
gammas <- tf$Variable(gammas, dtype = "float32")
beta <- tf$constant(c(1, -1), dtype = "float32")
eta <- tf$constant(c(-1, -2), dtype = "float32")
y_train <- matrix(c(0, 0, 0, 1, 0, 1, 0, 0, 0, 0), nrow = 2, byrow = TRUE)
y_train <- tf$constant(y_train, dtype = "float32")
compute_logLik(gammas = gammas, beta = beta, eta = eta, y_train = y_train)

gammas <- matrix(c(-1, -1, -1, -1), nrow = 1, byrow = TRUE)
gammas <- tf$Variable(gammas, dtype = "float32")
beta <- tf$constant(c(1, -1), dtype = "float32")
eta <- tf$constant(c(-1, -2), dtype = "float32")
y_train <- matrix(c(0, 0, 0, 1, 0, 1, 0, 0, 0, 0), nrow = 2, byrow = TRUE)
y_train <- tf$constant(y_train, dtype = "float32")
compute_logLik(gammas = gammas, beta = beta, eta = eta, y_train = y_train)
}
