% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_experiments.R
\name{eval_batchwise}
\alias{eval_batchwise}
\title{Evaluate fitted ontram models}
\usage{
eval_batchwise(model, x, y, im, bs, target)
}
\description{
Evaluate fitted ontram models
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
x_valid <- x_train[1:20,]
y_valid <- y_train[1:20,]
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train),
    method = "logit", n_batches = 10, epochs = 10)
mo_hist <- fit_ontram(mo, x_train = x_train, y_train = y_train, history = TRUE,
    x_test = x_valid, y_test = y_valid)
debugonce(eval_batchwise)
out <- eval_batchwise(mo, x_valid, y_valid, NULL, 5, as.numeric(wine$rating))
loss_from_eval(out$preds)
plots_from_eval(out$preds)
}
