% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ontram.R
\name{fit_ontram3}
\alias{fit_ontram3}
\title{Function for estimating the model}
\usage{
fit_ontram3(
  model,
  history = FALSE,
  x_train = NULL,
  y_train,
  img_train = NULL,
  save_model = FALSE,
  x_test = NULL,
  y_test = NULL,
  img_test = NULL,
  lambda2 = 1e-04,
  numnet = 1
)
}
\description{
Function for estimating the model
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
x_valid <- x_train[1:20,]
y_valid <- y_train[1:20,]
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train),
    method = "logit", n_batches = 10, epochs = 50)
mo_hist <- fit_ontram2(mo, x_train = x_train, y_train = y_train, history = TRUE,
    x_test = x_valid, y_test = y_valid)
plot(mo_hist)
}
