% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights_ontram}
\alias{get_weights_ontram}
\title{Get weights}
\usage{
get_weights_ontram(model, w_baseline = FALSE, w_shift = FALSE, w_image = FALSE)
}
\arguments{
\item{model}{an object of class \code{\link{ontram}}.}

\item{w_baseline}{logical. Whether weights of baseline model should be extracted.}

\item{w_shift}{logical. Whether weights of shift model should be extracted.}

\item{w_image}{logical. Whether weights of image model should be extracted.}
}
\description{
Get weights
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train),
                  method = "logit", n_batches = 10, epochs = 50)
fit_ontram(mo, x_train = x_train, y_train = y_train)
weights <- get_weights_ontram(mo, w_baseline = FALSE, w_shift = TRUE, w_image = FALSE)
}
