% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{k_mod_baseline_reparametrized}
\alias{k_mod_baseline_reparametrized}
\title{Reparametrized keras mbl}
\usage{
k_mod_baseline_reparametrized(K, thetas, ...)
}
\description{
Reparametrized keras mbl
}
\details{
\eqn{\nu_k = \nu_k' + \epsilon}
}
\examples{
library(tram)
data("wine", package = "ordinal")
mod_polr <- Polr(rating ~ 0, data = wine)
thetas <- coef(mod_polr, with_baseline = TRUE)
m <- k_mod_baseline_reparametrized(5L, thetas)
get_weights(m)
}
