% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{k_ontram_acc}
\alias{k_ontram_acc}
\title{Accuracy function}
\usage{
k_ontram_acc(K)
}
\description{
Accuracy function
}
\examples{
k_acc <- k_ontram_acc(ncol(Y))
k_acc(k_constant(Y), m(list(INT, X, Z)))
k_acc(k_constant(Y), m$output)
}
