% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram_metrics.R
\name{k_ontram_auc}
\alias{k_ontram_auc}
\title{AUC function}
\usage{
k_ontram_auc(K, cutoff = 3L)
}
\description{
AUC function
}
\examples{
k_auc <- k_ontram_auc(ncol(Y))
k_auc(k_constant(Y), m(list(INT, X, Z)))
k_auc(k_constant(Y), m$output)
}
