% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{k_ontram_binll}
\alias{k_ontram_binll}
\title{Binary logLik function}
\usage{
k_ontram_binll(K, cutoff = 3L)
}
\description{
Binary logLik function
}
\examples{
k_binll <- k_ontram_binll(ncol(Y))
debugonce(k_binll)
k_binll(k_constant(Y), m(list(INT, X, Z)))
k_binll(k_constant(Y), m$output)
}
