% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram_metrics.R
\name{k_ontram_loss}
\alias{k_ontram_loss}
\title{Another keras implementation of the ontram loss}
\usage{
k_ontram_loss(K)
}
\description{
Another keras implementation of the ontram loss
}
\examples{
y_true <- k_constant(matrix(c(1, 0, 0, 0, 0), nrow = 1L))
loss <- k_ontram_loss(ncol(y_true))
loss(y_true, m$output)
loss(k_constant(Y), m(list(INT, X, Z)))
}
