% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{k_ontram_qwk}
\alias{k_ontram_qwk}
\title{Continuous qwk}
\usage{
k_ontram_qwk(K, p = 2L)
}
\description{
Continuous qwk
}
\examples{
k_qwk <- k_ontram_qwk(ncol(Y))
k_qwk(k_constant(Y), m(list(INT, X, Z)))
k_qwk(k_constant(Y), m$output)
}
