% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{k_ontram_rps}
\alias{k_ontram_rps}
\title{CRPS loss}
\usage{
k_ontram_rps(K)

k_ontram_rps(K)
}
\description{
CRPS loss

RPS loss
}
\examples{
rps_loss <- k_ontram_rps(ncol(Y))
rps_loss <- k_ontram_rps(ncol(Y))
}
