% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{layer_trafo_intercept}
\alias{layer_trafo_intercept}
\title{Layer for transforming raw intercepts using softplus function}
\usage{
layer_trafo_intercept()
}
\description{
Layer for transforming raw intercepts using softplus function
}
