% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{metric_ontram_crossent}
\alias{metric_ontram_crossent}
\title{Categorical cross-entropy metric}
\usage{
metric_ontram_crossent(K)
}
\description{
Categorical cross-entropy metric
}
\examples{
cent <- metric_ontram_crossent(ncol(Y))
cent(y_true, m$output)
cent(k_constant(Y), m(list(INT, X, Z)))
}
