% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{metric_rps}
\alias{metric_rps}
\title{CRPS metric}
\usage{
metric_rps(K)
}
\description{
CRPS metric
}
\examples{
k_rps <- metric_rps(ncol(Y))
k_rps(k_constant(Y), m$output)
k_rps(k_constant(Y), m(list(INT, X, Z)))
}
