% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_baseline.R
\name{mod_baseline}
\alias{mod_baseline}
\title{Sequential model for the intercept function}
\usage{
mod_baseline(y_dim, ...)
}
\value{
Returns a model of type \code{\link[keras]{keras_mod_sequential}}
}
\description{
Fully connected neural network with input dimension K and
    output dimension K - 1
}
\examples{
m <- mod_baseline(5L)
}
