% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shift.R
\name{mod_shift}
\alias{mod_shift}
\title{Sequential model for the shift function}
\usage{
mod_shift(x_dim)
}
\value{
Returns a model of type \code{\link[keras]{keras_mod_sequential}}
}
\description{
Fully connected neural network with input dimension p and
    output dimension 1
}
\examples{
m <- mod_shift(5L)
}
