% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontram.R
\name{ontram}
\alias{ontram}
\title{General ordinal transformation neural network}
\usage{
ontram(
  mod_bl,
  mod_sh,
  mod_im = NULL,
  y_dim,
  x_dim,
  img_dim = NULL,
  method = c("logit", "cloglog", "loglog", "probit"),
  n_batches = 16,
  epochs = 50,
  lr = 0.001,
  optimizer = tf$keras$optimizers$Adam(learning_rate = lr),
  response_varying = FALSE
)
}
\description{
General ordinal transformation neural network
}
