% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontram.R
\name{ontram_polr}
\alias{ontram_polr}
\title{Combine intercept and tabular model}
\usage{
ontram_polr(
  x_dim,
  y_dim,
  method = c("logit", "cloglog", "loglog", "probit"),
  n_batches = 16,
  epochs = 50,
  lr = 0.001,
  optimizer = tf$keras$optimizer$Adam(learning_rate = lr)
)
}
\description{
Simple proportional odds logistic regression neural network,
    without response-varying effects and a simple linear predictor for the
    explanatory covariates.
}
\examples{
ontram_polr(2L, 5L)
}
