% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ontram_history}
\alias{plot.ontram_history}
\title{Plot ontram history}
\usage{
\method{plot}{ontram_history}(object, col_train = "blue", col_test = "red", add_best = FALSE, ...)
}
\description{
Plot ontram history
}
