% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcm-k_ontram.R
\name{plot.pbcm}
\alias{plot.pbcm}
\title{Plot PBCM}
\usage{
\method{plot}{pbcm}(
  object,
  obs_av = FALSE,
  titles = paste0("Sampled~from~Model~", 1:dim(object[[1]])[3]),
  labels = paste0("Mod ", 1:dim(object[[1]])[3]),
  cols = 1:dim(object[[1]])[3],
  n_col = NULL,
  n_row = NULL
)
}
\arguments{
\item{object}{output of function \code{\link{pbcm}}.}

\item{obs_av}{logical. Whether to plot mean of observed NLL as line (instead of ECDF).}

\item{titles}{titles used for facets.}

\item{labels}{labels used for fitted models.}

\item{cols}{colors.}

\item{n_col, n_row}{number of columns and rows for displaying plots.}
}
\description{
Plot PBCM
}
