% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{predict.k_ontram}
\alias{predict.k_ontram}
\title{S3 methods for \code{k_ontram}}
\usage{
\method{predict}{k_ontram}(
  object,
  x,
  type = c("distribution", "density", "trafo", "baseline_only"),
  ...
)
}
\description{
S3 methods for \code{k_ontram}
}
