% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ontram}
\alias{predict.ontram}
\title{Function for predicting cdf, pdf and class and compute logLik}
\usage{
\method{predict}{ontram}(model, x = NULL, y, im = NULL)
}
\description{
Function for predicting cdf, pdf and class and compute logLik
}
