% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ontram_rv}
\alias{predict.ontram_rv}
\title{Function for predicting cdf, pdf and class and compute logLik}
\usage{
\method{predict}{ontram_rv}(model, x, y, im = NULL)
}
\description{
Function for predicting cdf, pdf and class and compute logLik
}
