% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-load_history.R
\name{save_history_ontram}
\alias{save_history_ontram}
\alias{save_ontram_history}
\alias{load_ontram_history}
\title{Save/Load ontram history}
\usage{
save_ontram_history(object, filepath)

load_ontram_history(filepath)
}
\description{
Save/Load ontram history
}
