% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{simulate.k_ontram}
\alias{simulate.k_ontram}
\title{Simulate Responses}
\usage{
\method{simulate}{k_ontram}(object, x, nsim = 1, levels = NULL, seed = NULL)
}
\arguments{
\item{object}{an object of class \code{\link{k_ontram}}.}

\item{x}{list of data matrices (including matrix containing 1 if model intercept is non-complex)}

\item{nsim}{number of simulations.}

\item{levels}{levels of simulated ordered responses.}

\item{seed}{random seed.}
}
\description{
Simulate Responses
}
\examples{
data(wine, package = "ordinal")
fm <- rating ~ temp + contact
y <- model.matrix(~ 0 + rating, data = wine)
x <- ontram:::.rm_int(model.matrix(fm, data = wine))
loss <- k_ontram_loss(ncol(y))

mbl <- k_mod_baseline(ncol(y), name = "baseline")
msh <- mod_shift(ncol(x), name = "linear_shift")

mo <- k_ontram(mbl, msh)
compile(mo, optimizer = optimizer_adam(learning_rate = 10^-4), loss = loss)
fit(mo, x = list(matrix(1, nrow = nrow(wine)), x), y = y, batch_size = nrow(wine), epoch = 10)
simulate(mo, x = list(matrix(1, nrow = nrow(wine)), x), nsim = 1)
}
