% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.ontram}
\alias{simulate.ontram}
\title{Simulate Responses}
\usage{
\method{simulate}{ontram}(
  object,
  nsim = 1,
  x = NULL,
  y,
  im = NULL,
  levels = 1:ncol(y),
  seed = NULL,
  ...
)
}
\description{
Simulate Responses
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
x_valid <- x_train[1:20,]
y_valid <- y_train[1:20,]
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train),
                  method = "logit", n_batches = 10, epochs = 50)
fit_ontram(mo, x_train = x_train, y_train = y_train)
simulate(mo, nsim = 1, x = x_valid, y = y_valid)
}
