% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{warm_start.ontram_rv}
\alias{warm_start.ontram_rv}
\title{Set initial weights}
\usage{
\method{warm_start}{ontram_rv}(model, model_w, which = c("all", "baseline only", "shift only"))
}
\arguments{
\item{model}{an object of class \code{ontram_rv}.}

\item{model_w}{an object of class \code{\link[tram]{tram}}, \code{ontram_rv} or \code{\link{ontram}}
from which model weights are taken.}
}
\description{
Set initial weights
}
\examples{
library(tram)
set.seed(2021)
data("wine", package = "ordinal")
wine$noise <- rnorm(nrow(wine), sd = 0.3) + as.numeric(wine$rating)
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L, drop = FALSE]
y_train <- model.matrix(~ 0 + rating, data = wine)
im_train <- model.matrix(rating ~ noise, data = wine)[, -1L, drop = FALSE]

mp <- Polr(fml, data = wine)
mbl1 <- keras_model_sequential() \%>\%
         layer_dense(units = 4, input_shape = 1L, activation = "tanh") \%>\%
         layer_dense(ncol(y_train) - 1)
msh1 <- mod_shift(ncol(x_train))
mo_rv1 <- ontram(mod_bl = mbl1, mod_sh = msh1, method = "logit",
                 x_dim = ncol(x_train), y_dim = ncol(y_train),
                 response_varying = TRUE)
fit_ontram(mo_rv1, x_train = x_train, y_train = y_train, img_train = im_train)

mbl2 <- keras_model_sequential() \%>\%
         layer_dense(units = 4, input_shape = 1L, activation = "tanh") \%>\%
         layer_dense(ncol(y_train) - 1)
msh2 <- mod_shift(ncol(x_train))
mo_rv2 <- ontram(mod_bl = mbl2, mod_sh = msh2, method = "logit",
                 x_dim = ncol(x_train), y_dim = ncol(y_train),
                 response_varying = TRUE)
mod_weights(mo_rv2)
warm_start(mo_rv2, mo_rv1, "baseline only")
mod_weights(mo_rv2)
coef(mp)
warm_start(mo_rv2, mp, "shift only")
mod_weights(mo_rv2)
}
