% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{warmstart.k_ontram}
\alias{warmstart.k_ontram}
\title{Set initial weights}
\usage{
\method{warmstart}{k_ontram}(object, object_w, which = c("all", "baseline only", "shift only"))
}
\arguments{
\item{object}{an object of class \code{\link{k_ontram}}.}

\item{object_w}{an object of class \code{\link[tram]{Polr}} or \code{\link{k_ontram}} from which weights are taken.}
}
\description{
Set initial weights
}
\section{Notes}{

\itemize{
\item{For models of class \code{k_ontram_rv} the biases of the last layer are set to the intercepts of the \code{\link[tram]{Polr}} model
if \code{which} is set to "baseline_only" or "all".}
\item{The warmstarted model has to be assigned to a new variable.}
}
}

\examples{
library(tram)
set.seed(2021)
data(wine, package = "ordinal")
wine$noise <- rnorm(nrow(wine))
y <- model.matrix(~ 0 + rating, data = wine)
x <- ontram:::.rm_int(model.matrix(rating ~ temp + contact, data = wine))
im <- ontram:::.rm_int(model.matrix(rating ~ noise, data = wine))
loss <- k_ontram_loss(ncol(y))
mod_polr <- Polr(rating ~ temp + contact, data = wine)
mbl <- k_mod_baseline(ncol(y), name = "baseline")
msh <- mod_shift(ncol(x), name = "shift")
mim <- keras_model_sequential() \%>\%
  layer_dense(units = 8, input_shape = 1L, activation = "relu") \%>\%
  layer_dense(units = 16, activation = "relu") \%>\%
  layer_dense(units = 1, use_bias = FALSE)
mo <- k_ontram(mbl, list(msh, mim))
mo <- warmstart(mo, mod_polr, which = "baseline only") # object: k_ontram, object_w = Polr

compile(mo, loss = loss, optimizer = optimizer_adam(learning_rate = 1e-2))
fit(mo, x = list(matrix(1, nrow = nrow(wine)), x, im), y = y)
mbl2 <- k_mod_baseline(ncol(y))
mim2 <- keras_model_sequential() \%>\%
  layer_dense(units = 8, input_shape = 1L, activation = "relu") \%>\%
  layer_dense(units = 16, activation = "relu") \%>\%
  layer_dense(units = 1, use_bias = FALSE)
mo2 <- k_ontram(mbl2, mim2)
mo2 <- warmstart(mo2, mo, which = "all")  # object: k_ontram, object_w = k_ontram

mbl3 <- keras_model_sequential() \%>\%
  layer_dense(units = 8, input_shape = 1L, activation = "relu") \%>\%
  layer_dense(units = 16, activation = "relu") \%>\%
  layer_dense(units = ncol(y) - 1, use_bias = FALSE) \%>\%
  layer_trafo_intercept()()
mo3 <- k_ontram(mbl3, msh, response_varying = TRUE)
mo3 <- warmstart(mo3, mod_polr, "all") # object: k_ontram_rv, object_w = Polr
}
