% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-ontram.R
\name{warmstart.k_ontram_ci}
\alias{warmstart.k_ontram_ci}
\title{Set initial weights}
\usage{
\method{warmstart}{k_ontram_ci}(
  object,
  thetas = NULL,
  betas = NULL,
  which = c("all", "baseline only", "shift only")
)
}
\arguments{
\item{thetas}{intercepts of a \code{\link[tram]{Polr}} model as a vector (are added to the last layer during training).}

\item{betas}{shift terms of a \code{\link[tram]{Polr}} model as a vector (used to initialize the weights of the simple shift model).}
}
\description{
Set initial weights
}
\section{IMPORTANT}{

\itemize{
\item{The warmstarted model has to be assigned to a new variable.}
}
}

\examples{
library(tram)
set.seed(2021)
data(wine, package = "ordinal")
wine$noise <- rnorm(nrow(wine))
y <- model.matrix(~ 0 + rating, data = wine)
x <- ontram:::.rm_int(model.matrix(rating ~ temp + contact, data = wine))
im <- ontram:::.rm_int(model.matrix(rating ~ noise, data = wine))
loss <- k_ontram_loss(ncol(y))
mbl <- keras_model_sequential() \%>\%
  layer_dense(units = 8, input_shape = 1L, activation = "relu") \%>\%
  layer_dense(units = 16, activation = "relu") \%>\%
  layer_dense(units = ncol(y) - 1, use_bias = FALSE) \%>\%
  layer_trafo_intercept()()
mo <- k_ontram(mbl, msh, complex_intercept = TRUE)
mo <- warmstart(mo, thetas = coef(mod_polr, with_baseline = TRUE)[1:4L],
                betas = coef(mod_polr), which = "all")
}
