% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{warmstart.ontram}
\alias{warmstart.ontram}
\title{Set initial weights}
\usage{
\method{warmstart}{ontram}(object, object_w, which = c("all", "baseline only", "shift only"))
}
\arguments{
\item{object}{an object of class \code{\link{ontram}}.}

\item{object_w}{an object of class \code{\link[tram]{Polr}} from which model weights are taken.}
}
\description{
Set initial weights
}
\examples{
library(tram)
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
mp <- Polr(fml, data = wine)
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train))
coef(mp, with_baseline = T)
warmstart(mo, mp, "all")
mod_weights(mo)
}
