% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights.ontram}
\alias{weights.ontram}
\title{Extract model weights}
\usage{
\method{weights}{ontram}(model)
}
\description{
Extract model weights
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train))
weights(mo)
}
