% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ontram.R
\name{fit_ontram}
\alias{fit_ontram}
\title{Function for estimating the model}
\usage{
fit_ontram(
  model,
  history = FALSE,
  x_train = NULL,
  y_train,
  img_train = NULL,
  x_test = NULL,
  y_test = NULL,
  img_test = NULL,
  early_stopping = FALSE,
  patience = 1,
  min_delta = 0,
  stop_train = TRUE,
  filepath = NULL
)
}
\arguments{
\item{model}{an object of class "\link[ontram]{ontram}".}

\item{history}{logical. If TRUE train and test are returned as list.}

\item{x_train}{tabular data used for training the model.}

\item{y_train}{response data (one-hot encoded) used for training the model.}

\item{img_train}{image data used for training the model.}

\item{x_test}{tabular data used for evaluating the model.}

\item{y_test}{response data (one-hot encoded) used for evaluating the model.}

\item{img_test}{image data used for evaluating the model.}

\item{early_stopping}{logical. Whether to use early stopping (requires \code{history = TRUE}).}

\item{patience}{number of epochs with no improvement after which training will be stopped.}

\item{min_delta}{minimum increase in test loss considered as no improvement.}

\item{stop_train}{logical. Whether to stop training if conditions are fulfilled for the first time or whether to continue training.}

\item{filepath}{path where to save best model if \code{save_best = TRUE}.}

\item{save_best}{logical. Whether best model should be saved as HDF file.}
}
\description{
Function for estimating the model
}
\examples{
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
x_valid <- x_train[1:20,]
y_valid <- y_train[1:20,]
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train),
    method = "logit", n_batches = 10, epochs = 50)
mo_hist <- fit_ontram(mo, x_train = x_train, y_train = y_train, history = TRUE,
    x_test = x_valid, y_test = y_valid)
plot(mo_hist)
}
