% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{warm_start.ontram}
\alias{warm_start.ontram}
\title{Set initial weights}
\usage{
\method{warm_start}{ontram}(model, model_w, which = c("all", "baseline only", "shift only"))
}
\arguments{
\item{model}{an object of class \code{\link{ontram}}.}

\item{model_w}{an object of class \code{\link[tram]{tram}} or \code{\link{ontram}} from which weights are taken.}
}
\description{
Set initial weights
}
\examples{
library(tram)
data("wine", package = "ordinal")
fml <- rating ~ temp + contact
x_train <- model.matrix(fml, data = wine)[, -1L]
y_train <- model.matrix(~ 0 + rating, data = wine)
mp <- Polr(fml, data = wine)
mo <- ontram_polr(x_dim = ncol(x_train), y_dim = ncol(y_train))
coef(mp, with_baseline = T)
warm_start(mo, mp, "all")
weights(mo)
}
